Instead of finding NOUN the ADJ, just find ADJ
Phrases of the form "Noun the Adjective" are really, really uncommon on
twitter, but adjectives and nouns preceded by "the" are common.

We now pull these adjectives and nouns and use them to generate a name. The main
part of the name is hardcoded to "Prince Leopold", but he might be "Prince
Leopold of the Corn" or "Prince Leopold the Speedy".